rename: /components/menu-bar-icons → /components/menu-icons
