+++ components/menu-icons/MonitoringIcon.tsx
