+++ components/menu-icons/SalesIcon.tsx
